fixing tmux deployment

+++ dotdep_tmux.sh
@@ -1,4 +1,11 @@
-mkdir -p ~/.tmux/plugins/
-ln -s ~/repositories/tpm ~/.tmux/plugins/
-ln -s ~/repositories/dotfiles/.tmux.conf ~/
+if [ ! -d ~/.tmux/plugins/tpm ]
+then 
+	mkdir -p ~/.tmux/plugins/
+	ln -s ~/repositories/tpm ~/.tmux/plugins/
+fi
+if [ ! -f ~/.tmux.conf ]
+then
+	ln -s ~/repositories/dotfiles/.tmux.conf ~/
+fi
+mkdir -p ~/.config/systemd/user/ #should be fixed to not require this in future versions of tmux
 tmux source ~/.tmux.conf
